The register reaches one, and the lint job that had been red for three commits
Two things, and the second is the one I should have caught first.

THE LINT JOB WAS RED, and had been since 76da7cb9. Go CI run 2426 says:

    internal/governance/workflows.go:141:14: unnecessary conversion (unconvert)
        if n > int64(math.MaxInt64/int64(perUnit)) {

The inner conversion already yields int64. One character of dead syntax in
the duration-ceiling check, and it failed the Lint job on three consecutive
heads while I reported each of them verified.

The cause is my loop, not the code. build, vet, the whole test tree and the
52 shell guards all passed, every time — and none of them is golangci-lint,
which CI runs as its own job. The linter was installed here all along; the
version was too old to load a go1.26 module's config, which is exactly the
kind of failure that reads as "not available" and gets skipped. Installing
v2.13.2, the version ci.yml pins, took one command. It then found a second
issue in the very batch I was about to push: formatDuration in
internal/common/health, orphaned by deleting the third HealthService. Both
are fixed and the tree is at 0 issues.

THE REGISTER: TEN TO ONE.

internal/backup: LocalStorage and the Storage interface it implemented, the
progress-tracking Reader and Progress, the DirEntry/FileInfo shims and the
whole os-function indirection layer the tests substituted. Every one of them
existed so LocalStorage could walk a directory, and nothing constructed
LocalStorage. This one was holding up a false claim in a document an
operator reads during a disaster: PRODUCTION-READINESS said Manager "routes
through the Storage interface; both LocalStorage and S3Storage are wired".
It does not. Local backups are os.ReadFile / os.WriteFile against
StorageDir; S3 is reached through the concrete *S3Storage. The capability
the sentence was really claiming is real — a created backup is uploaded, a
restore falls back to the bucket — so the document is corrected rather than
the code bent to match a design nobody adopted.

internal/risk: AlertManager (dedup, severity routing and an email nothing
sends; detection and the security_alerts row were never the gap),
BehaviorTracker (a second set of behavioural baselines;
admin/continuous_auth.go computes the one that runs), DeviceFingerprinter
(canvas/WebGL capture no console page performs; the live device signal is
the posture agent), and RiskAssessment's three methods — GetSignalSummary,
ToJSON, GetHighRiskSignals — a reporting surface no caller used, whose only
exercise was three tests written against it.

internal/audit.AnomalyDetector: brute-force detection over audit events,
constructed by nothing, duplicating internal/risk.

Three live helpers came out of those files rather than dying with them:
haversineDistance (the impossible-travel detector and two signals measure
with it), TrustLevel (the scorer's device-trust signal), IsPrivateIP (its
IP-reputation signal). Each has a file and a sentence saying why it stayed.

AND A GUARD THAT WOULD HAVE GONE GREEN BY CHECKING NOTHING.
TestSecurityAlertWritersNameOnlyRealColumns read two files by name — the
two I deleted. A hard-coded list would have passed on an empty set, while
internal/risk/anomaly.go, the writer that actually runs, was never on it.
It now finds its writers by searching, and fails if it finds none.

Left: internal/access.UpstreamPool, whose verdict is a CRUD surface or a
migration dropping two tables. That is its own piece of work.

Verified: build + vet + the whole Go tree, golangci-lint v2.13.2 (the
version ci.yml pins) at 0 issues, guards 52/52, deadservice 10 -> 1.

+++ internal/governance/workflows.go
@@ -138,7 +138,7 @@ func parseDuration(s string) (time.Duration, error) {
 	}
 	// A window this long is a mistake or an attack, and letting it through is
 	// how an int64 of nanoseconds wraps to a date in the past.
-	if n > int64(math.MaxInt64/int64(perUnit)) {
+	if n > math.MaxInt64/int64(perUnit) {
 		return 0, fmt.Errorf("invalid duration %q: that is longer than any access window can be", s)
 	}
 	return time.Duration(n) * perUnit, nil
